Plugin structure for the boys
