+++ .gitignore
@@ -1,3 +1,4 @@
+.DS_Store
 refresh_token.txt
 __pycache__/
 runtime/
