+++ roboronya/plugins/chucknorris.py
@@ -1,0 +1,28 @@
+description = 'Look for a ~~random joke~~ undeniable truth about our lord and savior.'
+
+from roboronya.plugins.plugin import *
+
+class Command(Plugin):
+
+    def run(roboronya, conv, cmd_args, **kwargs):
+        random_joke = requests.get(
+            config.CHUCK_API_URL
+        ).json()
+        if random_joke.get('type') == 'success':
+            return roboronya.send_message(
+                conv, random_joke['value']['joke'], **kwargs
+            )
+        logger.info(
+            '{} Failed to retrieve joke from {}. '
+            'Got response: {}'.format(
+                kwargs['log_tag'], config.CHUCK_API_URL, random_joke,
+            )
+        )
+        return roboronya.send_message(
+            conv,
+            (
+                'Sorry I could not find a joke '
+                'Please try again later.'
+            ),
+            **kwargs
+        )
